apps/api/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
SQLAlchemy>=2.0
httpx>=0.27
python-dotenv>=1.0
websockets>=12.0
claude-code-sdk>=0.0.20
mcp[cli]>=1.14.1
uv>=0.4.20
docker>=7.0
cryptography>=42.0
openai>=1.40
unidiff>=0.7
aiohttp>=3.9
rich>=13.0
python-multipart>=0.0.6
itsdangerous>=2.2
stripe>=10.0
google-auth>=2.29
PyYAML>=6.0
